implementing data laoder for pnet
